+++ tests/test_core/test_hooks.py
@@ -22,7 +22,7 @@ async def test_must_be_rendering_in_layout_to_use_hooks():
     with pytest.raises(RuntimeError, match="No life cycle hook is active"):
         await SimpleComponentWithHook().render()
 
-    with idom.Layout(SimpleComponentWithHook()) as layout:
+    async with idom.Layout(SimpleComponentWithHook()) as layout:
         await layout.render()
 
 
@@ -35,7 +35,7 @@ async def test_simple_stateful_component():
 
     sse = SimpleStatefulComponent()
 
-    with idom.Layout(sse) as layout:
+    async with idom.Layout(sse) as layout:
         patch_1 = await render_json_patch(layout)
         assert patch_1.path == ""
         assert_same_items(
@@ -75,7 +75,7 @@ async def test_set_state_callback_identity_is_preserved():
 
     sse = SimpleStatefulComponent()
 
-    with idom.Layout(sse) as layout:
+    async with idom.Layout(sse) as layout:
         await layout.render()
         await layout.render()
         await layout.render()
@@ -108,7 +108,7 @@ async def test_use_state_with_constructor():
         state, set_inner_state.current = idom.use_state(make_default)
         return idom.html.div(state)
 
-    with idom.Layout(Outer()) as layout:
+    async with idom.Layout(Outer()) as layout:
         await layout.render()
 
         assert constructor_call_count.current == 1
@@ -141,7 +141,7 @@ async def test_set_state_with_reducer_instead_of_value():
         count.current, set_count.current = idom.hooks.use_state(0)
         return idom.html.div(count.current)
 
-    with idom.Layout(Counter()) as layout:
+    async with idom.Layout(Counter()) as layout:
         await layout.render()
 
         for i in range(4):
@@ -308,7 +308,7 @@ async def test_use_effect_callback_occurs_after_full_render_is_complete():
         effect_triggers_after_final_render.current = not effect_triggered.current
         return idom.html.div()
 
-    with idom.Layout(OuterComponent()) as layout:
+    async with idom.Layout(OuterComponent()) as layout:
         await layout.render()
 
     assert effect_triggered.current
@@ -336,7 +336,7 @@ async def test_use_effect_cleanup_occurs_before_next_effect():
 
         return idom.html.div()
 
-    with idom.Layout(ComponentWithEffect()) as layout:
+    async with idom.Layout(ComponentWithEffect()) as layout:
         await layout.render()
 
         assert not cleanup_triggered.current
@@ -375,7 +375,7 @@ async def test_use_effect_cleanup_occurs_on_will_unmount():
 
         return idom.html.div()
 
-    with idom.Layout(OuterComponent()) as layout:
+    async with idom.Layout(OuterComponent()) as layout:
         await layout.render()
 
         assert not cleanup_triggered.current
@@ -406,7 +406,7 @@ async def test_memoized_effect_on_recreated_if_dependencies_change():
 
         return idom.html.div()
 
-    with idom.Layout(ComponentWithMemoizedEffect()) as layout:
+    async with idom.Layout(ComponentWithMemoizedEffect()) as layout:
         await layout.render()
 
         assert effect_run_count.current == 1
@@ -449,7 +449,7 @@ async def test_memoized_effect_cleanup_only_triggered_before_new_effect():
 
         return idom.html.div()
 
-    with idom.Layout(ComponentWithEffect()) as layout:
+    async with idom.Layout(ComponentWithEffect()) as layout:
         await layout.render()
 
         assert cleanup_trigger_count.current == 0
@@ -476,7 +476,7 @@ async def test_use_async_effect():
 
         return idom.html.div()
 
-    with idom.Layout(ComponentWithAsyncEffect()) as layout:
+    async with idom.Layout(ComponentWithAsyncEffect()) as layout:
         await layout.render()
         await asyncio.wait_for(effect_ran.wait(), 1)
 
@@ -496,7 +496,7 @@ async def test_use_async_effect_cleanup():
 
         return idom.html.div()
 
-    with idom.Layout(ComponentWithAsyncEffect()) as layout:
+    async with idom.Layout(ComponentWithAsyncEffect()) as layout:
         await layout.render()
 
         component_hook.latest.schedule_render()
@@ -527,7 +527,7 @@ async def test_use_async_effect_cancel(caplog):
 
         return idom.html.div()
 
-    with idom.Layout(ComponentWithLongWaitingEffect()) as layout:
+    async with idom.Layout(ComponentWithLongWaitingEffect()) as layout:
         await layout.render()
 
         await effect_ran.wait()
@@ -554,7 +554,7 @@ async def test_error_in_effect_is_gracefully_handled(caplog):
         return idom.html.div()
 
     with assert_idom_logged(match_message=r"Layout post-render effect .* failed"):
-        with idom.Layout(ComponentWithEffect()) as layout:
+        async with idom.Layout(ComponentWithEffect()) as layout:
             await layout.render()  # no error
 
 
@@ -575,7 +575,7 @@ async def test_error_in_effect_cleanup_is_gracefully_handled(caplog):
         return idom.html.div()
 
     with assert_idom_logged(match_error=r"Layout post-render effect .* failed"):
-        with idom.Layout(ComponentWithEffect()) as layout:
+        async with idom.Layout(ComponentWithEffect()) as layout:
             await layout.render()
             component_hook.latest.schedule_render()
             await layout.render()  # no error
@@ -604,7 +604,7 @@ async def test_error_in_effect_pre_unmount_cleanup_is_gracefully_handled():
         match_message=r"Pre-unmount effect .*? failed",
         error_type=ValueError,
     ):
-        with idom.Layout(OuterComponent()) as layout:
+        async with idom.Layout(OuterComponent()) as layout:
             await layout.render()
             set_key.current("second")
             await layout.render()  # no error
@@ -629,7 +629,7 @@ async def test_use_reducer():
         )
         return idom.html.div()
 
-    with idom.Layout(Counter(0)) as layout:
+    async with idom.Layout(Counter(0)) as layout:
         await layout.render()
 
         assert saved_count.current == 0
@@ -659,7 +659,7 @@ async def test_use_reducer_dispatch_callback_identity_is_preserved():
         saved_dispatchers.append(idom.hooks.use_reducer(reducer, 0)[1])
         return idom.html.div()
 
-    with idom.Layout(ComponentWithUseReduce()) as layout:
+    async with idom.Layout(ComponentWithUseReduce()) as layout:
         for _ in range(3):
             await layout.render()
             saved_dispatchers[-1]("increment")
@@ -679,7 +679,7 @@ async def test_use_callback_identity():
         used_callbacks.append(idom.hooks.use_callback(lambda: None))
         return idom.html.div()
 
-    with idom.Layout(ComponentWithRef()) as layout:
+    async with idom.Layout(ComponentWithRef()) as layout:
         await layout.render()
         component_hook.latest.schedule_render()
         await layout.render()
@@ -705,7 +705,7 @@ async def test_use_callback_memoization():
         used_callbacks.append(cb)
         return idom.html.div()
 
-    with idom.Layout(ComponentWithRef()) as layout:
+    async with idom.Layout(ComponentWithRef()) as layout:
         await layout.render()
         set_state_hook.current(1)
         await layout.render()
@@ -733,7 +733,7 @@ async def test_use_memo():
         used_values.append(value)
         return idom.html.div()
 
-    with idom.Layout(ComponentWithMemo()) as layout:
+    async with idom.Layout(ComponentWithMemo()) as layout:
         await layout.render()
         set_state_hook.current(1)
         await layout.render()
@@ -758,7 +758,7 @@ async def test_use_memo_always_runs_if_dependencies_are_none():
         used_values.append(value)
         return idom.html.div()
 
-    with idom.Layout(ComponentWithMemo()) as layout:
+    async with idom.Layout(ComponentWithMemo()) as layout:
         await layout.render()
         component_hook.latest.schedule_render()
         await layout.render()
@@ -785,7 +785,7 @@ async def test_use_memo_with_stored_deps_is_empty_tuple_after_deps_are_none():
         used_values.append(value)
         return idom.html.div()
 
-    with idom.Layout(ComponentWithMemo()) as layout:
+    async with idom.Layout(ComponentWithMemo()) as layout:
         await layout.render()
         component_hook.latest.schedule_render()
         deps_used_in_memo.current = None
@@ -810,7 +810,7 @@ async def test_use_memo_never_runs_if_deps_is_empty_list():
         used_values.append(value)
         return idom.html.div()
 
-    with idom.Layout(ComponentWithMemo()) as layout:
+    async with idom.Layout(ComponentWithMemo()) as layout:
         await layout.render()
         component_hook.latest.schedule_render()
         await layout.render()
@@ -830,7 +830,7 @@ async def test_use_ref():
         used_refs.append(idom.hooks.use_ref(1))
         return idom.html.div()
 
-    with idom.Layout(ComponentWithRef()) as layout:
+    async with idom.Layout(ComponentWithRef()) as layout:
         await layout.render()
         component_hook.latest.schedule_render()
         await layout.render()
@@ -865,7 +865,7 @@ async def test_use_effect_automatically_infers_closure_values():
 
         return idom.html.div()
 
-    with idom.Layout(CounterWithEffect()) as layout:
+    async with idom.Layout(CounterWithEffect()) as layout:
         await layout.render()
         await did_effect.wait()
         did_effect.clear()
@@ -893,7 +893,7 @@ async def test_use_memo_automatically_infers_closure_values():
 
         return idom.html.div()
 
-    with idom.Layout(CounterWithEffect()) as layout:
+    async with idom.Layout(CounterWithEffect()) as layout:
         await layout.render()
         await did_memo.wait()
         did_memo.clear()
@@ -918,7 +918,7 @@ async def test_use_context_default_value():
         value.current = idom.use_context(Context)
         return html.div()
 
-    with idom.Layout(ComponentProvidesContext()) as layout:
+    async with idom.Layout(ComponentProvidesContext()) as layout:
         await layout.render()
         assert value.current == "something"
 
@@ -927,7 +927,7 @@ async def test_use_context_default_value():
         value.current = idom.use_context(Context)
         return html.div()
 
-    with idom.Layout(ComponentUsesContext()) as layout:
+    async with idom.Layout(ComponentUsesContext()) as layout:
         await layout.render()
         assert value.current == "something"
 
@@ -958,7 +958,7 @@ async def test_use_context_only_renders_for_value_change():
         render_count.current += 1
         return html.div()
 
-    with idom.Layout(ComponentProvidesContext()) as layout:
+    async with idom.Layout(ComponentProvidesContext()) as layout:
         await layout.render()
         assert render_count.current == 1
 
@@ -995,7 +995,7 @@ async def test_use_context_updates_components_even_if_memoized():
         render_count.current += 1
         return html.div()
 
-    with idom.Layout(ComponentProvidesContext()) as layout:
+    async with idom.Layout(ComponentProvidesContext()) as layout:
         await layout.render()
         assert render_count.current == 1
         assert value.current == 0
@@ -1041,7 +1041,7 @@ async def test_nested_contexts_do_not_conflict():
         inner_render_count.current += 1
         return html.div()
 
-    with idom.Layout(Root()) as layout:
+    async with idom.Layout(Root()) as layout:
         await layout.render()
         assert outer_render_count.current == 1
         assert inner_render_count.current == 1
@@ -1097,7 +1097,7 @@ async def test_neighboring_contexts_do_not_conflict():
         right_used_value.current = idom.use_context(RightContext)
         return idom.html.div()
 
-    with idom.Layout(Root()) as layout:
+    async with idom.Layout(Root()) as layout:
         await layout.render()
         assert left_render_count.current == 1
         assert right_render_count.current == 1
@@ -1142,7 +1142,7 @@ async def test_error_in_effect_cleanup_is_gracefully_handled():
         error_type=ValueError,
         match_error="The error message",
     ):
-        with idom.Layout(ComponentWithEffect()) as layout:
+        async with idom.Layout(ComponentWithEffect()) as layout:
             await layout.render()
             component_hook.latest.schedule_render()
             await layout.render()  # no error
@@ -1159,7 +1159,7 @@ async def test_set_state_during_render():
             set_state(state + 1)
         return html.div(state)
 
-    with Layout(SetStateDuringRender()) as layout:
+    async with Layout(SetStateDuringRender()) as layout:
         await layout.render()
         assert render_count.current == 1
         await layout.render()