+++ docs/source/about/changelog.rst
@@ -444,7 +444,7 @@ See :ref:`Custom JavaScript Components` for details on the new interface.
 - Make docs section margins larger - :issue:`450`
 - Search broken in docs - :issue:`443`
 - Move src/idom/client out of Python package - :issue:`429`
-- Use composition instead of classes with Layout and LifeCycleHook  - :issue:`412`
+- Use composition instead of classes async with Layout and LifeCycleHook  - :issue:`412`
 - Remove Python language extension - :issue:`282`
 - Add keys to models so React doesn't complain of child arrays requiring them -
   :issue:`255`